backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-cors-headers==4.3.1
openai==1.3.0
anthropic==0.7.0
google-generativeai==0.3.1
sentence-transformers==2.2.2
numpy==1.24.3
scikit-learn==1.3.2
